requirements.txt
coinbase
reedsolo
python-bitcoinrpc
pycoin

